core: Make bear dependencies instance bound
Bears can now define dependencies at instantiation for `BEAR_DEPS`,
depending on different section settings.

The Bear class makes now a copy of the dependencies specified in
`BEAR_DEPS`, so users can directly modify the set inside `__init__`
(which prevents modifying the class wide dependencies that are always
needed).

For that, the core test bears had to be adapted, because before
picklibility of those test bears was not necessary to run the
core.

Closes https://github.com/coala/coala/issues/4560